Implement Main and fix issues

+++ Userservice/src/main/java/com/userservice/entity/User.java
@@ -2,17 +2,15 @@ package com.userservice.entity;
 
 import lombok.Data;
 import jakarta.persistence.*;
-
 import java.time.LocalDateTime;
 
 @Data
 @Entity
 @Table(name = "users", schema = "public")
 public class User {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(nullable = false, length = 100)
     private String name;
